feat: 개발 끝, codegen 리팩토링

+++ src/components/common/layout/header/Header.container.tsx
@@ -1,58 +1,62 @@
-import { useMutation } from "@apollo/client";
 import React, { useState } from "react";
 import { useNavigate } from "react-router";
-import {
-  Mutation,
-  MutationDeleteTokenArgs,
-} from "../../../../commons/types/generated/types";
-import { DELETE_TOKEN } from "../../../login/Login.query";
+import { useDeleteTokenMutation } from "../../../../commons/graphql/generated";
 import HeaderPresenter from "./Header.presenter";
 
 const HeaderContainer = (props: any) => {
   const navigate = useNavigate();
+  //* 검색창 오픈
   const [isSearch, setIsSearch] = useState(false);
+
+  //* 검색어 상태
   const [search, setSearch] = useState("");
 
-  const handleSearchWord = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setSearch(() => e.target.value);
-  };
-
-  const [deleteToken] = useMutation<Mutation, MutationDeleteTokenArgs>(
-    DELETE_TOKEN
-  );
+  //* 토큰 삭제
+  const [deleteToken] = useDeleteTokenMutation();
 
   const tokenId = sessionStorage.getItem("token");
+  const userId = sessionStorage.getItem("userId");
 
+  //* 로그아웃
   const handleLogout = async () => {
-    try {
-      const { data } = await deleteToken({
-        variables: {
-          input: {
-            where: {
-              id: String(tokenId),
+    if (userId) {
+      try {
+        const { data } = await deleteToken({
+          variables: {
+            input: {
+              where: {
+                id: String(tokenId),
+              },
             },
           },
-        },
-      });
-      if (!data?.deleteToken?.token) {
-        return;
+        });
+        sessionStorage.clear();
+        alert("로그아웃 됐습니다.");
+        navigate("/");
+        if (!data?.deleteToken?.token) {
+          return;
+        }
+      } catch (e) {
+        alert("오류가 발생했습니다.");
       }
-      sessionStorage.clear();
-      alert("로그아웃 됐습니다.");
-      navigate("/");
-    } catch (e) {
-      console.log(e);
+    } else {
+      navigate("/login");
     }
+  };
+
+  //* 검색어
+  const handleSearchWord = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setSearch(() => e.target.value);
   };
   return (
     <>
       <HeaderPresenter
-        isSearch={isSearch}
-        setIsSearch={setIsSearch}
-        handleSearchWord={handleSearchWord}
         search={search}
+        isSearch={isSearch}
         setSearch={setSearch}
+        setIsSearch={setIsSearch}
         handleLogout={handleLogout}
+        handleSearchWord={handleSearchWord}
       />
     </>
   );
